Move src/__tests__ to tests
